// src/generated/models/PlayerEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hnefatafl API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PlayerEnum = {
    None: 'None',
    Viking: 'Viking',
    Monster: 'Monster',
    Both: 'Both'
} as const;
export type PlayerEnum = typeof PlayerEnum[keyof typeof PlayerEnum];


export function PlayerEnumFromJSON(json: any): PlayerEnum {
    return PlayerEnumFromJSONTyped(json, false);
}

export function PlayerEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlayerEnum {
    return json as PlayerEnum;
}

export function PlayerEnumToJSON(value?: PlayerEnum | null): any {
    return value as any;
}